+++ src/main/kotlin/com/atlassian/performance/tools/report/api/FullReport.kt
@@ -1,5 +1,6 @@
 package com.atlassian.performance.tools.report.api
 
+import com.atlassian.performance.tools.io.api.ensureDirectory
 import com.atlassian.performance.tools.jiraactions.api.ActionMetricStatistics
 import com.atlassian.performance.tools.report.api.action.EditedIssuesReport
 import com.atlassian.performance.tools.report.api.action.SearchJqlReport
@@ -58,6 +59,11 @@ class FullReport {
             SearchJqlReport(
                 allMetrics = actionMetrics
             ).report(cohortWorkspace)
+
+            WaterfallHighlightReport().report(
+                metrics = actionMetrics,
+                workspace = TestWorkspace(cohortWorkspace.resolve("WaterfallHighlight").ensureDirectory())
+            )
         }
     }
 }